+++ VultisigApp/VultisigApp/Model/Folder.swift
@@ -22,12 +22,11 @@ class Folder: Hashable, Equatable {
     }
     
     static func == (lhs: Folder, rhs: Folder) -> Bool {
-        lhs.folderName == rhs.folderName && lhs.containedVaultNames == rhs.containedVaultNames
+        lhs.id == rhs.id
     }
     
     func hash(into hasher: inout Hasher) {
-        hasher.combine(folderName)
-        hasher.combine(containedVaultNames)
+        hasher.combine(id)
     }
     
     static let example = Folder(folderName: "Folder", containedVaultNames: ["12345"], order: 0)
